make parallel test a task
